Changed minimum value of API polling frequency
The previous minimum value of the Last.FM API polling frequency was set to 10 seconds. In retrospect, that value was a little conservative and I think that it's too slow to react quickly enough to new songs being played.

I'll also investigate some measures to make sure we're not polling the API when music is not playing in the future.